Launch optionally with isolated classpath (support for Flux)
User has to add --function.runner.isolated=true on the command line.
In this case the app is launched in an isolated classpath with
reactor in the parent class loader. Then the function is launched
in a class loader with the same parent so it can support functions
of Flux.