Fix Java stubs so they map better to the package names used
Before this change there is a java package declared in the .proto
files in the java build, but it doesn't get applied to the .proto
files defined in the parent, leading to a slightly weird split
of classes between io.projectriff..* and top level function and
fntypes packages.

It seems like the only way to do this is to preprocess the source
files. Maven isn't the best way to do that, but at least it's only
a few lines.