Consume metrics and use them for auto-scaling
Avoid scaling down to zero when the input topic is non-empty. This is
necessary because production and consumption rates could be zero for a while
and yet the input topic could contain arbitrarily many unprocessed messages.

The autoscaler unit tests resort to white box testing to avoid races. Rather
than running receive and sampling loops, the autoscaler is told when to:
  * receive specific metrics, and
  * take a sample.