Approximate current scaling behaviour with restructured autoscaler
The autoscaler consumes metrics and usess them in a minimal way simply to
detect activity on an input topic and thereby defer scaling to zero. The meat
of the scaling algorithm is based on queue length which approximates the
current offset-based algorithm. The hope is to modify the algorithm in the
future to depend more on metrics and a little less on queue length.

The autoscaler is a collection of decorated metrics scalers indexed by
function id. Each metrics scaler proposes a number of replicas (0 or 1 - see
above) based on metrics. The decorations:
  * propose replicas proportional to the queue length of the input topic,
  * limit scaling up according to configuration,
  * smooth the number of replicas over time,
  * limit scaling down according to input topic queue length, and
  * delay scaling to zero for a configurable time period.