Show 404 upon message to non-existent topic
Previously, riff would pass messages forward to Kafka, would create a
Kafka topic automatically. However, this Kakfa topic is not visible as
a Riff topic managed by Kubernetes. The expected behaviour was instead
to return a 404.

This commit introduces the concept of riff topics into the
http-gateway. The gateway now checks to see if a topic exists before
sending a message to the message transport (currently, Kafka).

- Introduce TopicHelper
- Check for TopicExists() during /message/ handling

[Github projectriff/riff#410]

Fixes #410