Improve topic 404 support
Distinguish b/w Kubernetes 404s and other errors. Previously, when looking up
Riff topics in Kubernetes, it was assumed that any error returned was
essentially a 404. These changes distinguish between:

1) Finding a topic
2) Not finding the topic
3) Other errors

Also in this commit is an ugly-looking set of changes to the
requests_handler_test.go file. This is because of a race condition in test
configuration -- without breaking apart the requests handler test into
separate Contexts with repeated initialisation, the test falsely fails on an
unrelated error.

Improve the helper interface:
  * Document interface and constructor function.
  * Rename TopicHelper to RiffTopicExistenceChecker.
  * Return err from NewTopicHelper instead of panic
  * Introduce parameter for namespace in TopicExists.
  * Introduce Clientset param in NewRiffTopicExistenceChecker to simplify the interface.