Avoid defaulting the invoker version
When using cobra to set the default value for the --invoker-version flag
for `riff init <invoker>` and `riff create <invoker>` the wrong default
version can be selected at runtime.

Previously, we used the list of invokers to dynamically install sub
commands for `init` and `create`. The default for the invoker version
was set based on the invoker version. With a set of invokers installed
like:

```
NAME      VERSION
command   0.0.6-snapshot
go        0.0.2-snapshot
java      0.0.6-snapshot
node      0.0.6-snapshot
python2   0.0.6-snapshot
python3   0.0.6-snapshot
```

When creating a function with the go invoker, cobra is defaulting the
.InvokerVersion to `0.0.6-snapshot`, instead of the `0.0.2-snapshot`.
When the go invoker is the only invoker install, it defaults correctly
to `0.0.2-snapshot`.

We can work arround this issue by not defaulting the .InvokerVersion in
cobra, but by manually defaulting the value.