Better error messages for riff create
When running `riff create` without an invoker (or an invoker that is not
installed), we know the command will fail. The build command will later
attempt to convert the first argument to the filepath param, which will
be validated against the current directory. By disabling all other
validation for the create command, we can provide a proper error
message (which is generated from the init command).

Issue #555