Add backoff loop
Really unsophisticated solution to #323, just backs off for 1
second and makes the RPC call again. If the function invoker dies
(or is replaced with a telepresence) then the sidecar will
retry until it succeeds.

...unless the command line flag --exitOnComplete=true, in which case
it simply exits when the RPC call completes. This is super useful for
event source and sink use cases (Consumer and Supplier in Java terms).