Expose function sidecar protocol to the inovker
There are two different protocols the function sidecar can use to
communicate with the function invoker: http and grcp. Previously, the
invoker had no way to know how it would be invoked, so the invoker
needed to setup a server to receive traffic for each protocol.

The configured function sidecar to invoker protocol is now set in the
function invoker pod as the RIFF_FUNCTION_INVOKER_PROTOCOL env var.

Refs projectriff/java-function-invoker#73