Ignore unknown falgs for init and create commands
If the user attempts to run a command like:

    riff create java --handler functions.Uppercase

...but they do not have the java invoker installed, they currently
recieve an unhelpful error saying the `--handler` flag is unknown. Since
the init and create root comamnds require an invoker, and some invokers
require the handler flag while others do no, we can relax the
enforcement of unknown flags for these commands. Subcommands continue to
enforce unknown flags.

The ability to supress unknown flag warnings is new in pflags v1.0.1.

Fixes #555