Remove err return from TopicExists() signature
In the previous implementation, performing a Get() to fetch information
about a riff topic could result in an error -- eg if the remote
controller is failed or unreachable. This was surfaced to calling code.

This was slightly ugly and hard to reason about for callers, especially
when the alwaysTrueTopicExistenceChecker was added and broke the
semantic model of "it might return an error".

The updated implementation uses a shared informer. One practical effect
is that there is no direct error to pass back to callers.

As with the commit that introduced err, this has a highly disruptive
effect on requests_handler_test.go due to a change in indentation.

[projectriff/riff#553]