Remove vestigial stub existence checker
Previously this type was used to test the behaviour of the message and
request handlers if TopicExists() returned an error as its second value.
In a previous commit, the signature of TopicExists() was updated to
remove the error value. Tests for the error case were removed, but the
errorRiffTopicExistenceChecker stub was overlooked.