Pass stop channel into TopicInformer
Previously we had an unusued stop channel provided to the informer Run()
to satisfy its interface, but did not use it to signal the end of
processing.

In this change, the http-gateway main() creates and injects a stop
channel, as it is the main() function that also closes the HTTP server
when a shutdown signal is received.