Repair 404 performance regression by using a shared TopicInformer (#558)
Squashed commit, see #558 for detailed history. Fixes #553.

* Replace topic Get() logic with a TopicInformer

The previous implementation of the RiffTopicExistenceChecker would
directly query the controller process for every single message. This
worked but has severe effects on performance.

This commit instead uses a k8s shared informer to keep track of riff
topics being added or removed.

* Remove err return from TopicExists() signature

In the previous implementation, performing a Get() to fetch information
about a riff topic could result in an error -- eg if the remote
controller is failed or unreachable. This was surfaced to calling code.

This was slightly ugly and hard to reason about for callers, especially
when the alwaysTrueTopicExistenceChecker was added and broke the
semantic model of "it might return an error".

The updated implementation uses a shared informer. One practical effect
is that there is no direct error to pass back to callers.

As with the commit that introduced err, this has a highly disruptive
effect on requests_handler_test.go due to a change in indentation.

* Log key-lookup errors

Instead of ignoring errors, we at least log them. A panic seems like
overkill.

Some reformatting.

* Add mutex to improve thread-safety of knownTopics

* Format empty structs to single lines

* Move mutex decl/init above other variables

Stylistic change, intended to show the scope of variables the mutex is
intended to control.

* Remove vestigial stub existence checker

Previously this type was used to test the behaviour of the message and
request handlers if TopicExists() returned an error as its second value.
In a previous commit, the signature of TopicExists() was updated to
remove the error value. Tests for the error case were removed, but the
errorRiffTopicExistenceChecker stub was overlooked.

* Return from AddFunc/DeleteFunc for err'd keys

* Make AddFunc/DeleteFunc logs more robust & general

- Remove %s for err-producing key value, as the key is likely to be a
  junk value
- Switch error format from %+v to %#v (I mix these two up frequently)
- Remove mention of "map" from success log line; this reveals an
  implementation detail

* Add TODOs for KeyFunc implementations.

See the discussion at https://github.com/projectriff/riff/pull/558#discussion_r184437224

These TODOs should be addressed as part of projectriff/riff#529 and
projectriff/riff#485.

* Move mutex to emphasise it controls knownTopics

* Pass stop channel into TopicInformer

Previously we had an unusued stop channel provided to the informer Run()
to satisfy its interface, but did not use it to signal the end of
processing.

In this change, the http-gateway main() creates and injects a stop
channel, as it is the main() function that also closes the HTTP server
when a shutdown signal is received.

* Use single stop channel

Since we are using close() to signal, we do not need separate channels
for each of the http-gateway Run() and the existence checker Run().

* Use defer close(done) idiom for stop channel

See https://blog.golang.org/pipelines for a discussion.