Check for topic in all namespaces
This makes the behaviour more similar to what it was before we started
checking for the existence of the topic.

Keep track of namespaces per topic and log warning if duplicates
created/removed.

Add comments to topic-controller's README explaining some of the impact of the
current behaviour where topic resources are namespaced while the broker topics
are cluster scoped.

Notes:

- riff currently only supports deploying a function or topic with a specific
  name into a single namespace; deploying the same resource with the same name
  to two different namespaces is not supported.

- The same function and topic can be created in two namespaces, but which one
  will be used is undefined.

- If a topic that exists in two namespaces is removed from one of the
  namespaces, the gateway will not be able to find the topic from the other
  namespace.

See #485

Fixes #529