Parameterize ports between invoker and sidecar
The port used for the function sidecar to communicate with the function
invoker is currently shared knowledge. Each invoker needs to know to
listen on 8080 for http or 10382 for grpc. This knowledge is hard coded
into the sidecar and each function invoker.

Instead the ports should be determined by the function-controller and
set into the function's deployment. For the sidecar there is a new
`--port` argument, and for the invoker there are environment variables
HTTP_PORT and GRPC_PORT.

There are two port variables for the invoker because there are cases
where an invoker may need to support http and grpc.

These ports should no longer be hard coded in invokers. Although the
values have not changed.