Remove corrlationId header from http protocol
Function invokers using the http protocol do not need to be concerned
with the correlationId. It's confusing to invoker implementors because
they think they do need to handle and return this value.

The sidecar will set the correlationId on the http response from the
invoker, so the invoker has no influence over the value. It should not
be sent to the invoker, and any value from the invoker should be
ignored.

Fixes #531