Support stream gateway in function controller
Support is provided by enabling a feature flag BEFORE the function controller is
deployed. The feature flag is enabled by setting the property
stream-gateway=enabled in the riff-feature-flags configmap in the riff-system
namespace.

For example, the feature flag may be enabled as follows:

kubectl create configmap riff-feature-flags --from-literal=stream-gateway=enabled --namespace=riff-system

When the feature flag is enabled:
* A service is created for each link which fronts the gRPC servers of the link's
  function pods
* The function pod consists of the main container only (with no function sidecar)