Divorce scenario logic from receiver logic
Previously the stubReceiver acted as a crossroads between the
SimulationUpdater and MetricsReceiver interface. That is, the
communication point between the simulator and the autoscaler. But this
means that providing multiple scenarios is essentially impossible.
Instead, until now, there has been a single "CombinedScenario".

This commit breaks apart the roles so that stubReceiver no longer fills
both roles. Instead SimulationUpdater methods are updated to accept a
MetricsReceiver as a parameter. This opens the door to breaking apart
scenarios.

Also included is an extraction of "common" scenario logic into
scenario.go.