Provide feedback from commands
Feedback is provided from commands which would not otherwise provide feedback.

Note that the feedback is only that the command itself is successful, since
resources are often created asynchronously in knative/kubernetes.

Currently the feedback message is generic, for example:

$ riff channel create newchan --bus stub
riff channel create completed successfully

More specific feedback, for example "channel newchan created successfully",
can be added in the future if necessary.

Closes https://github.com/pivotal-cf-experimental/riff-cli/issues/63