Inject CLI config for manifests, invokers and buildpacks
Manifests, invokers and buildpacks are configuration that is built into
the riff cli. It should be easy to create a custom build of riff with a
different configuration. By moving the config into a central location
and then passing that configuration into other component, it will be
easier to customize in the future.

Fixes #912