Parameterize 'riff' name
It should be eaiser to interact with the riff cli when it has a
different name. Most instances of "riff" that would appear on the cli
have been parameterized with a value that can be defined at build time.

riff-distro has also been updated. Where there was no difference between
riff and riff-distro commands the distro command has been replaced with
the riff command.

This change attempted to preserve the current docs directory content,
unless the content was incompatible with a parameterized world. For
example, the listing of invokers/buildpacks is too bespoke to properly
express.

Refs: #912