Restore correct manifests versions
The manifests were moved in #931 and at the same time regressed to older
buggy versions of Knative. At the same time, the moved manifests were
referenced directly in commands ignoring the values injected into the
commands. For the purposes of OSL tracking, this cannot happen. The
manifest defintitions need to remain inside the main package and not be
directly referenced by other portions of the codebase.

The riff-distro main now creates it's own default set of manifests and
injects those values into the riff-distro codebase. For simplicity, I
removed all other manifests except for 'latest', since keeping two sets
in sync will be difficult. In the long run, riff-distro users should
provide am explicit path to a manifest rather than use the compiled in
versions.