Drop usage of the invoker *arg*
The invoker *arg* is gone, and its previous meaning too.
It is now an optional `--invoker` flag, which is used to override the
detection done by the buildpack. It does not drive selection of an
invoker URL anymore.

The --invoker-url and <custom> support is also removed

Adds support for latest BP build for --local-path

Fixes #945