Ensure there are pods in the istio-system namespace
If pods are slow to be created there may not immediately be pods that
exist to check if they are running. Previous, this would pass the test.
Now we require there be at least three pods in the namespace.
Additionally, we wait 10 seconds before the first test, rather than
testing immediately.