Use a cluster BuildTemplate
A ClusterBuildTemplate has the advantage of being defined once in the
cluster and can be updated and maintained with the rest of riff.

The cache PVC has moved into a separate resource as it and a service
account are still required in each namespace in order to run builds.

Fixes #968