Set default image prefix durring namespace init
We can often infer based on the credentials used durring namespace init
what the default image path should be.

For GCR, the project id can be parsed out of the token and prepended
with 'gcr.io'. For Docker Hub the username is passed directly can is
prepended with 'docker.io'.