Default image for function create based on namespace init credentials (#1105)
During`riff function create` if `--image` is not defined, a default value is provided based on the name of the function and a value (default-image-prefix) looked up from a ConfigMap (riff-build) in the target namespace. If the ConfigMap doesn't exist or the prefix is not present in the map, then --image will be required.

The ConfigMap is created during namespace init if it doesn't exist, otherwise it is updated. The value is automatically derived from the credentials being set for Docker Hub and GCR. Defining the secret out of band or using generic basic auth defined in #1103 provide insufficient metadata to set a prefix. For other auth strategies or cases where the inference may be wrong, the prefix can be defined during `riff namespace init` via the `--image-prefix` flag.

Docker Hub: `docker.io/{username}`
GCR: `gcr.io/{project-id}`