URL Cleanup
This commit updates URLs to prefer the https protocol. Redirects are not followed to avoid accidentally expanding intentionally shortened URLs (i.e. if using a URL shortener).

# Fixed URLs

## Fixed But Review Recommended
These URLs were fixed, but the https status was not OK. However, the https status was the same as the http request or http redirected to an https URL, so they were migrated. Your review is recommended.

* http://%s/%s (UnknownHostException) with 4 occurrences migrated to:
  https://%s/%s ([https](https://%s/%s) result UnknownHostException).
* http://istio-release (UnknownHostException) with 4 occurrences migrated to:
  https://istio-release ([https](https://istio-release) result UnknownHostException).

## Fixed Success
These URLs were switched to an https URL with a 2xx status. While the status was successful, your review is still recommended.

* http://example.com with 2 occurrences migrated to:
  https://example.com ([https](https://example.com) result 200).
* http://golang.org/pkg/os/ with 1 occurrences migrated to:
  https://golang.org/pkg/os/ ([https](https://golang.org/pkg/os/) result 200).
* http://projectriff.io with 2 occurrences migrated to:
  https://projectriff.io ([https](https://projectriff.io) result 200).

# Ignored
These URLs were intentionally ignored.

* http://localhost:12345/nope.yaml with 1 occurrences