Allow riff.toml to exist
Function authors may define and commit their own riff.toml file which
will be used for both local and on-cluster builds. The riff cli allows
users to override values in the riff.toml when creating a function.

packs now allows environment variables to be passed to the detect and
build phases. The riff cli no longer writes its own riff.toml instead
defering to the user's riff.toml and instead passes it's custom config
as env vars. Function buildpacks understand both riff.toml and the env
var, favoring the environment when there is a conflict.

Refs #1192