Apply the Istio patch to istio.yaml
The istio.yaml published as part of the Knative Serving release is too
heavy in both the number of replicas and reservations. We previously
scaled it back using a "patch" yaml that duplicates the resources that
needed to change. Now we use a single yaml file with the patch applied.

The diff from the Knative version is also published.