Avoid Actions for state mutations
The Action API is far more robust than we need. We can simplying what
test authors need to specify for updates and deletes by avoiding the
actions and focusing on the core essense of what is changing.

For updates it is the mutated resource. For deletes it is a reference to
the resource being deleted.