Set credentials implementation (#1270)
The command will now create and update credentials to contain
basic auth secrets for registires. The credential details are passed as
arguments with passwords read via stdin, or a prompt on the command
line. DockerHub, GCR and basic auth are supported.

Optionally set default image prefix when setting credentials
Note: this capability is only available for DockerHub and GCR as basic
auth cannot adequately define an image prefix.