Introduce Executable interface for options
Each options object now contains everything it needs to run the command
without the cobra.Command. The actual cobra.Command calls into the
options.

It is now possible to unit test the comamnd behavior without cobra,
although it is still recomended to test through cobra since this is how
the options are actually used at runtime.