Add color to tables
Output is now colorized for table lists with the Format* functions doing
the right thing by default.

This nesessitated hacking the tabwriter since the color code sequences
were affecting the tab width. This was easy to integrate since the
caller is also vendored within our code bases, but will need to be
rethought when we switch to mainline code.