Local application and function builds (#1283)
For application and function create calls with --local-path, we use pack
to build and publish the container from local source before creating the
application or function resource.

The created application/function will not have a Source field in the
spec. The image is resolved to a digest by the riff system controller.

This PR uses mockery to mock the pack client for tests. Use of the mock
is limited to the few tests that use pack. We can swap to a different
mock/stub library in the future.

Co-Authored-By: Eric Bottard <313494+ericbottard@users.noreply.github.com>