Backport default image prefix support (#1292)
The --image flag is now optional for applications and functions. The
default value is `_`, which applies the configured default prefix
with the application/function's name. A custom repo or tag can be
specified like `_/repo:tag`.

For in-cluser builds this resolution occurs in the reconciler. For local
builds, it occurs within the CLI. In both cases, the resolution logic is
defined within projectriff/system.