Avoid deadlocks during create
Since we can’t rule out the case where either WaitUntilReady() or
ApplicationLogs() (or both) return an error just before the context expires,
the current code can deadlock. See this example flow in the case of both
failing: https://play.golang.org/p/76Dcr9VlAXo. A solution is to increase the
buffer size to 3 so that the main thread will never block on sending an error
to the channel.