RFC 0002 - Bindings (#1360)
Non-trivial workloads often need to access resources, like databases, message brokers or any credentialed resource. In Cloud Foundry, these are called services and they are provisioned from a broker and then bound to applications at both build and runtime. As Kubernetes has a different notion of a Service, this RFC will refer to "service bindings" as simply bindings.

The binding contains both metadata and credentials to provide a workload the details it needs to connect and consume the resource.

Co-Authored-By: Florent Biville <445792+fbiville@users.noreply.github.com>
Co-Authored-By: Swapnil Bawaskar <sbawaskar@pivotal.io>
Co-Authored-By: Eric Bottard <313494+ericbottard@users.noreply.github.com>
Co-Authored-By: Mark Fisher <markfisher@users.noreply.github.com>